// src/lib/types.ts
export interface CartItem {
  productId: string;
  variantId: string | null;
  name: string;
  price: number;
  quantity: number;
  imageUrl: string;
  taxRate?: number;
}

export interface Cart {
  items: CartItem[];
  couponCode?: string;
  discountAmount: number;
  taxAmount?: number;
  taxName?: string;
  taxInclusive?: boolean;
  subtotal: number;
  total: number;
  currency: string;
}